separated app declaration and keepers from app.go
